+++ app/src/renderer/components/staking/TabValidators.vue
@@ -6,7 +6,7 @@
       data-empty-search(v-else-if="!delegates.loading && sortedFilteredEnrichedDelegates.length === 0")
       template(v-else)
         panel-sort(:sort='sort', :properties="properties")
-        li-delegate(v-for='i in sortedFilteredEnrichedDelegates' :disabled="!userCanDelegate" :key='i.id' :delegate='i')
+        li-validator(v-for='i in sortedFilteredEnrichedDelegates' :disabled="!userCanDelegate" :key='i.id' :delegate='i')
 </template>
 
 <script>
@@ -14,7 +14,7 @@ import { mapGetters } from "vuex"
 import num from "scripts/num"
 import { includes, orderBy } from "lodash"
 import Mousetrap from "mousetrap"
-import LiDelegate from "staking/LiDelegate"
+import LiValidator from "staking/LiValidator"
 import { TmBtn, TmPage, TmDataEmpty, TmDataLoading } from "@tendermint/ui"
 import DataEmptySearch from "common/TmDataEmptySearch"
 import { calculateTokens } from "scripts/common"
@@ -24,7 +24,7 @@ import VmToolBar from "common/VmToolBar"
 export default {
   name: "page-staking",
   components: {
-    LiDelegate,
+    LiValidator,
     TmBtn,
     TmDataEmpty,
     DataEmptySearch,