+++ CHANGELOG.md
@@ -14,10 +14,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * added new validator component @okwme
 * simple loading indicator for page staking @faboweb
 * Created new "My Stake" tab. @NodeGuy
+* possibility to end unbonding transactions @faboweb
 
 ### Changed
 
 * Refactored Addressbook. @NodeGuy
+* Created new modal window for staking to a validator. @NodeGuy
+* Added new validator profile page @faboweb
+* cleaning up new validator profile page and balance header @jbibla
 
 ### Fixed
 
@@ -28,6 +32,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * delegations request would update state of next user after sign out sign in @ƒaboweb
 * enable user to stake only after the request for his current delegations has returned @ƒaboweb
 * cache undelegations @faboweb
+* show error instead of breaking Voyager if reconnection fails @faboweb
+* reconnection errors did not show up correctly in view @faboweb
 
 ## [0.10.2] - 2018-08-29